Rewrite singleton classes Catalog, Log, ResourceMap and TBB
Singletons are challenging in C++, it is very hard to have a thread-safe
implementation which takes into account dependencies between singletons.

Current code does not work as expected, we had issues on Windows when
loading an Excel plug-in because ResourceMap was not initialized.
Current code is difficult to read and debug, mostly because singleton
builders (*_init structs defined along singletons) are initialized in
header files and thus duplicated in each object file.

We propose here a simpler pattern: a MutexLockSingleton template class
is declared in MutexLock.hxx, it is modelled after ResourceMapInstance.
Catalog, Log, ResourceMap and TBB are rewritten to use it.

A static builder is defined in implementation files, and a second one in
GetInstance().  There are no threading issues because static variables
are instantiated during library load, before user code.  The second
builder ensures that initialization order does not matter.

It is not clear whether problems reported above are fixed by this commit,
but at least it should be simpler to follow singleton constructions.