Started changes in the parametrization of SpectralModel.
As the spectral model is the Fourier transform of the stationary covariance model, it
should have a similar interface compared to the covariance model.

Thus we add the following parameters to the SpectralModelImplementation class:
- scale parameter
- amplitude parameter
- Correlation parameter
- spatialDimension

To be the equivalent of a covariance model, scale should be of size spatial dimension, which is not necessary the underlying dimension.
Of course, the setters are also implemented.

The CauchyModel now uses the parameters of the parent class