Improved the numerical stability and accuracy of ClaytonCopula
This copula has a CDF/PDF and a random generator badly conditioned for large values of theta (the limit of this copula is the Min copula, which is singular), leading to overflows when implemented naively, and suffers from cancellation for theta close to zero (the limit being the independent copula). The previous implementation was not robust enough to handle these cases with full accuracy.