Fix Interval::intersect for mixed finite/infinite bounds
Method Interval::intersect did not take infinite bounds properly into
account and returned the min|max of the bounds regardless of the finite
flag.
It seems to impact RandomMixture and Distribution arithmetic
multiplication.

In cases both bounds from both interval are infinite, we set the
corresponding bound value to the numerical limit.
In truncated distribution class, we need to override
this behavior by manually setting bounds on distribution range values.