Improved GeneralizedLinearModelAlgorithm and KrigingAlgorithm
The main improvement is to delegate to LAPACK the solution of the least-squares problem for the trend computation instead of doing it by hand using a QR decomposition (it is exactly what LAPACK do, but faster).

The other improvements are cosmetic: better agreement with coding rules, renaming of protected methods and attributes.