Split function parameters value and description
This adds [gs]etParameterDescription
and modifies [gs]etParameter to return an NP instead of a
NPWithDescription for functions/evaluation/gradient/hessian classes.

This change make the API consistent with Distribution and
CovarianceModel.

It also fixes a problem with the overload of setParameter in several evaluation
implementation (discovered it with ComposedNMEI fot robopt module).
The evaluation implementation would have to overload both
setParameter(NP) and setParameter(NPWD) for the accessor to work.
I prefered to have a single accessor, and it should be a little faster
than copying NPWD.

Also some implementations of setParameter were missing in evaluation.