Fix a segfault in CovarianceModelImplementation::discretizeHMatrix
CovarianceModelImplementation::discretizeHMatrix contains this snippet:
  if (dimension_ == 1)
  {
    CovarianceAssemblyFunction simple(*this, vertices, nuggetFactor);
    covarianceHMatrix.assemble(simple, 'L');
  }

In fact, *this is converted into a local CovarianceModel variable,
and CovarianceAssemblyFunction::covarianceModel_ is a reference to
this local variable.  When exiting from CovarianceAssemblyFunction
constructor, this local variable is destroyed, but covarianceModel_
is still referencing it and covarianceHMatrix.assemble() crashes.

Also remove reference for vertices too, in order to be safe.