Improved the speed of the P1LagrangeEvaluationImplementation class.
Now the operator() method benefits from the const references returned by the Field and Mesh objects instead of costly copies. It greatly improves the speed of the evaluation over a point. The most demanding part of the evaluation over a sample is done in parallel, resulting in an additional speed-up. The cache mechanism is not activated in the evaluation over a sample.