Allow detection without bindings
The detection of OpenTURNS fails if swig dir is not there,
which is conditioned by the bindings being bult or not.
We may want to build a module without bindings.
If we want to build the module bindings although openturns bindings have been disabled
the build will fail because of missing swig files anyway.