Improved KarhunenLoeveP1Algorithm
Now, the Gram matrix is reused as is in the case of scalar covariance models.
Also fixed a bug: if the mesh has pending vertices (in particular if it has no simplices), then the Gram matrix does not have the correct dimension to make the algorithm work.
Also switched from P1Lagrange to PiecewiseLinear when spatial dimension is 1