Fix http://trac.openturns.org/ticket/875
TruncatedNormalFactory randomly crashes on osx systems. This is due to
the new implementation in the MaximumLikelihoodFactory class (reduced
objective function but not accurate bounds). Thus to fix the problem:

 - We update OptimizationProblem's bounds :  they should be zero size or
 should have the same size as objective's function input dimension
 Thus we redefine new bounds in case
 - We use OptimizationProblem::isValid to check Equality/Inequality
 constraints and we raise an exception if method returns false

Same work has been done on the MethodOfMoment class as we should get the
same problems if we face reduced objective functions