Improve performance of CovarianceModelImplementation::discretizeAndFactorize
When covariance model defines computeStandardRepresentative() method, the
covariance matrix is obtained as the Kronecker product of two matrices:
  C=Kron(A,B),
where A calls computeStandardRepresentative(), and B is the spatial
covariance, which is precomputed.
If we compute the Cholesky decomposition of A and B:
  A = La La^T
  B = Lb Lb^T
Then Lc = Kron(La,Lb).
As complexity of Cholesky decomposition is O(N^3), this method is
much faster in high dimension.

A protected member named definesComputeStandardRepresentative_ is added
into CovarianceModelImplementation, to tell if subclass defines
computeStandardRepresentative() method.
In this case, discretize() and discretizeAndFactorize() uses the
Kronecker product of matrices; otherwise current implement is unchanged,
by calling operator().

For now, discretizeRow has not been modified and always call operator().