Fix loading configuration file on OSX when installing with conda
Conda performs prefix substitution within libraries so that packages
are relocatable.  On OSX, line
  directory = String(SYSCONFIG_PATH) + PrefixConfigSubdirectory_;
causes trouble.  When compiling, conda makes SYSCONFIG_PATH a very long
string, and it is replaced during installation.
On Linux, it works as expected, string contains the new path, and /openturns
is appended to it afterwards.
On OSX, string also contains the new path, but it keeps its original length,
and thus /openturns is appended after lots of \0.

With
  directory = String() + String(SYSCONFIG_PATH).c_str() + PrefixConfigSubdirectory_;
SYSCONFIG_PATH is adjusted to have the new length.