Handle comments in BuildFromTextFile
PR #508 broke otpmml which relied on ImportFromTextFile to ignore comments.

Now lines starting with one of the chars contained in Sample-CommentMarkers setting are ignored.
This means the initial check for dimension has to look for the first data or description line.

This makes ImportFromTextFile closer to the behavior of numpy.loadtxt.

Also there was a small bug in ParseStringAsValues not eating blank spaces when the separator is ' '.