Fix chaos memory usage
Add accessors to the ProjectionStrategy interface to fix the reference counting in FunctionalChaosAlgorithm.
Fixes http://trac.openturns.org/ticket/927 (Functional chaos is memory hungry)

Note that now the projection strategy passed as argument of FCA is not modified in-place.
I kept the friendship of ProjectionStrategyImplementation to FCA only for the updateBasis method and not to copy the Proxy at that moment. But this remains a bad design.

I also think that we shouldnt build metamodel algorithms without passing the actual input/output data: in FCA we do lots of complicated stuff to override the data.