Disable multithreaded BLAS in GeneralLinearModelAlgorithm::maximizeReducedLogLikelihood
OpenBLAS performs badly when called from TBB.
In order to have the best performance, multithreading must be disabled
from high-level routines and not TBB::ParallelFor.

This is hackish, since we do not want to include cblas.h and/or omp.h just to change the
number of threads.  Add a CMake option DISCARD_BLAS_MULTITHREADING (ON by default) to
disable this change, in case it causes trouble on some systems.

If some computations run faster when OMP_NUM_THREADS environment variable is set to 1,
this means that TBB::{Disable,Enable}ThreadedBlas calls should be added to this method.