Mesh does no more derive from DomainImplementation
The aim is to let Mesh use the standard Interface/Implementation idiom.
This class is heavy, thus all methods returning a Mesh perform a copy
of an object which may be huge.

A MeshDomain class is added as an adaptor to see a Mesh as a
DomainImplementation.

The inheritance between Mesh and DomainImplementation was never used
in C++ code.  In Python, automatically convert a Mesh into a Domain
thanks to MeshDomain.