Modify Saltelli estimator
The numerator constant of the first order index was computed as mean of product of G(A), G(B) whereas it should be the product of means as recommended by saltelli02
This was not consistent with the documentation either where noted as G_0^2, so we explicited it also.
The variance is reduced resulting in a faster estimator.