New SymbolicFunction(Description,String,UnsignedInteger) constructor
This allows to define a single formula which has several return values.
For instance, the example given on ot-users can be written (in Python)
with ExprTk:

  inputs = ['Q', 'Ks', 'Zv', 'Zm', 'Hd', 'Zb', 'L', 'B']
  formula = """var alpha := (Zm - Zv)/L;
  var H := (Q/(Ks*B*sqrt(alpha)))^(3.0/5.0);
  var Zc := H + Zv;
  var Zd := Zb + Hd;
  var S := Zc - Zd;
  return [H,S]"""
  function = ot.SymbolicFunction(inputs, formula, 2)

This feature is also implemented with muParser, but it is much less powerful,
because muParser only concatenates independant statements.  For instance,
conversion between 2D cartesian and polar coordinates can be written as

  inputs = ['x', 'y']
  formula = """sqrt(x^2+y^2),
  (x == 0 && y == 0) ? 0 :
    (x < 0 && y == 0) ? - pi_  :
      2*atan(y/(x + sqrt(x^2 + y^2)))"""
  polar = ot.SymbolicFunction(inputs, formula, 2)