Speed-up evaluation of a PythonFunction on a Sample
Here is what happens when a PythonFunction is evaluated
on a Sample:

 1. If argument X is a 2d Python sequence, it is converted by SWIG
    into a Python Sample X1.  Otherwise X is a Python Sample
 2. PythonEvaluation::operator()(const Sample &inS) is called on the
    Sample wrapped by X (or X1).
 3. This method converts inS into a Python sequence P1.
 4. OpenTURNSPythonFunction.__call__(P1) is called
 5. It tries to convert P1 into a Point (with a copy), this
    fails and it converts P1 into a Python Sample S1 (with another
    copy).
 6. OpenTURNSPythonFunction._exec_sample(S1) is called and evaluated.

Dispatch at step 5 is already performed by PythonEvaluation::operator()
overloads, this part of this commit can be applied independantly.

An optional 'copy' argument is added to PythonFunction, to revert to old
code if something goes wrong.  Normally this should not happen, but who
knows?  It also allows to compute the cost of step 3.

Cache is removed, it prevents this optimization.

Rewrite default Function._exec_sample with list comprehension, this
is faster.