Improve KarhunenLoeveResult::project(ProcessSample)
Computations with TBB were buggy.
Rewrite project() to perform matrix-matrix multiplications without TBB,
this is much faster than TBB wih matrix-vector multiplications, and as
a bonus gives the right result.

Remove project(Basis), it is very likely that it was buggy too because
of TBB usage, and it was also very slow.