Changed the way UserDefinedCovarianceModel are built
Now, NonStationaryCovarianceModelFactory::buildAsUserDefinedCovarianceModel() is based on NonStationaryCovarianceModelFactory::buildAsCovarianceMatrix(), leading to a huge speed-up. This method allows to estimate the discretization of the covariance function of a ProcessSample over its mesh, using efficient linear algebra.
It is made possible thanks to a change the implementation of UserDefinedCovarianceModel: now, it is based on a big covariance matrix instead of many small blocks. It speeds up a lot the creation of the model and its discretization over its natural mesh.