New P1LagrangeInterpolation class
This class inherits from FieldFunctionImplementation,
and allow interpolating a Field on another Mesh.

This class is somewhat special, since it can deal with Fields
of any dimension (and output Field has the same dimension as
input Field), whereas input and output dimensions are usually
fixed by FieldFunction.
In order to provide basic checks during evaluation, a dimension
is declared in constructor, can be changed afterwards with the
setDimension() method.