Include internal OT directories first
A similar change had already been performed in 6f2986e to force SWIG
to include files from source directories and not installed ones.

But the same fix is also needed for lib when one sets CMAKE_INSTALL_PREFIX
to a path which is used by some system library.

This commit ensures that all local paths are searched before system paths.