Added the ExtremeValueCopula class
This class allows to define the most general 2D extreme value copulas through their Pickand function (A convex function A from [0,1] into [1/2, 1] such that max(t,1-t)<=A(t)<=1). The properties of the function are checked pointwise, see the documentation to know how to activate/deactivate the check and how to tune it.