Improved Kolmogorov test
Now, the p-value is computed using bootstrap when the reference distribution is estimated from the tested sample. See https://stats.stackexchange.com/questions/110272/a-naive-question-about-the-kolmogorov-smirnov-test for the details.
This improvement is not used in FunctionalChaosAlgorithm because it essentially leads to the same ranking of factories for a much larger execution time.