Add SobolSimulationAlgorithm
The algorithm allows to estimate Sobol indices by simulation.
It samples the distribution of a random vector X through a model f.
At each iteration a fixed number blockSize_ of replications inputs is generated (N = replicationSize_*(d+2))
These inputs are evaluated by blocks of size batchSize_ through the
model f.
Then we compute the distribution of the indices (first and total order) on this current replication sample.
We use one of the 4 available estimators (Saltelli, Jansen, ...)
At the end of each iteration we update the global distribution of the indices.

As stopping criteria, for each input component i, we check either:

- thightness: Phi_i^F(1-alpha)-Phi_i^F(alpha)<epsilon AND Phi_i^T(1-alpha)-Phi_i^T(alpha)<epsilon
- separability: Phi_i^F(1-alpha)<Phi_i^T(alpha)