Added a Robbins Monro based estimation of quantile in ProcessSample
This algorithm is described in https://projecteuclid.org/download/pdf_1/euclid.aoms/1177729586. It provides a consistent estimator of the quantile without the need to sort the data, and is much more memory-friendly than the empirical quantile estimation. It implements the nested procedure described in https://arxiv.org/abs/1510.00967. Note that the reference contains typos, not sure to have fixed all of them.
I let the basic implementation of the algorithm being visible to the user to allow debugging.

Cleaned ProcessSample

Several methods were put into dead ifdef clauses.