Fixed DirectionalSampling instability
When the origin is on the boundary of the event, it was impossible to make a difference between rays fully contained in the event and rays fully outside of the event, resulting in a wrong probability estimate. Now, the stability of the origin classification wrt the event is checked and the standard function is slightly modified if the origin is unstable.