Improved ProbabilitySimulationResult
Now the variance or the coefficient of variation are returned as a zero if the probability estimate is in (0, 1) and as not changed during at least two iterations. It allows to stop algorithms such as DirectionalSampling when each block gives the same contribution in (0, 1). This behavior can be deactivated through the 'ProbabilitySimulationResult' key of ResourceMap.