Fixed StudentFactory
The previous code was completely wrong. Now, the factory is able to estimate multivariate Student distributions, with the restriction of nu>2 due to the use of first and second moments for the location and scale parameters. The correlation is estimated in a robust way using the sample Kendall tau matrix, and the number of degrees of freedom nu is obtained by profiled log-likelihood maximization.