Improved the performance of MarginalTransformationEvaluation
The improvement of ResourceMap (direct access to a Scalar instead of a systematic conversion from a String) lead to an improvement of x1.94 wrt OT 1.11 on the benchmark (Uniform, Triangular)->(Normal, Normal). The introduction of a specific attribute to switch to tail evaluation gives an improvement of x2.48 wrt OT 1.11.