Improved UserDefined
Now, it uses the alias method for sampling, as described in Devroye86, Chapter 3. It has an optimal space and time complexity of resp. O(n) and O(1), instead of O(nlog(n)) and O(log(n)) with the previous implementation.
For a distribution with a support of size 10^4 the speed-up is of ~300%.