Added classes dedicated to calibration
They allow to compute the posterior distribution of the parameter of a given function in presence of noisy observations.
There are two kind of probabilistic contexts. The prior distribution of the parameter can be a Dirac distribution, in which case the observation error is also estimated (assuming a covariance error proportional to the identity matrix), or the prior distribution is given in terms of location and covariance through a normal distribution, in which case the covariance of the error is also given through a centered normal distribution.
There are also two kind of assumptions on the dependence of the model wrt its parameter: it can be linear or nonlinear.
The combinations of these contexts and assumptions lead to four methods:
+ LinearLeastSquaresCalibration for Dirac/linear dependence
+ NonLinearLeastSquaresCalibration for Dirac/nonlinear dependence
+ BLUE for Normal/linear dependence
+ ThreeDVAR for Normal/nonlinear dependence