Parallelize the evaluation of a Sample if n_cpus is specified.
Two cases to consider:

1. func_sample != None

In the master the evaluation on a sample with func_sample was never
parallel. With this commit if n_cpus is specified then the function
func_sample is parallelized with multiprocessing.

The use of multiprocessing causes an overhead. It is therefore
necessary to activate it only for functions that are expensive.
Up to user to check if there is a speed up.

2. func != None and func_sample == None

When introducing memoryview the multiprocessing has been disabled.
It is reactivated by this commit. As in case 1, if the calculation is
fast n_cpus should not be used.