Added reshape()/reshapeInPlace() methods to Matrix
It allows to present the data of a given matrix as a matrix of different dimensions. The reshape can be done in-place. There is no copy of data if the new dimensions lead to the same number of entries in the matrix, oterwise the data are truncated if the new flat size is less than the original flat size, or completed by zeros if the new flat size is greater than the original flat size.