Added the BlockIndependentDistribution
This work is based on the development of A. Dutfoy during the 2019 coding party. This distribution allows to create an n-dimensional distribution based on a collection of distributions of dimension n_1,...,n_k such that n=n_1+...+n_k and each distribution in the collection describes the behaviour of a group of variables independent from all the other variables.