Improved Distribution
Now it is possible to select between MonteCarlo or QuasiMonteCarlo to compute entropy for high dimension models. While QuasiMonteCarlo may be more accurate for a given sample size, it was taking too long to generate a sample for distributions with no dedicated implementation of the Rosenblatt transformation (eg the EmpiricalBernsetinCopula class).