Update tests & notebooks
First fix tests with new API. We drop as possible useless print. Basic
idea is to focus on surrogate quality (r2, max error). We rely on
ottesting for that.

Also we updated notebooks. For many examples illustrating beams, added a
`transformation` manually. Purpose is to have similar results are previously.
Also it shows how could we easily transform data & fit models with the
a more simple API