Fix estimate_multivariate_distribution example
FittingTest::BIC computes the loglikelihood.
It does not know whether it is dealing with a copula or any other multivariate distribution.
In order to correctly compute the BIC of a copula and resolve #1558, we need to transform
the considered sample because its 1D marginals need to agree with the
uniform distributon on [0,1].