Fix numpy broadcasting error in example
In `kriging_advanced.ipynb`, in order to draw confidence intervals, Samples are converted to numpy arrays of shape (1000,) with the np.hstack command and then added to a numpy array with shape (1,1000). The resulting array has shape (1000,1000). This is not the desired behavior.
It turns out that this conversion is useless, because the sum of a Sample with size 1000 and dimension 1 and of a numpy array with shape (1,1000) is a Sample with size 1000 and dimension 1 which can be drawn by matplotlib.