Fix the search box utility for install
The search utility is improved to narrow down the search to one of the following
categories :
- Examples
- API
- Theory

This is done in two steps.

First a custom scorer is implemented in the mysearchtools.js file. This file is
similar to the default file created by sphinx. The mysearchtools.js file is placed
in the _static/js directory.
This directory is copied through CMake install as sphinx seems unable to do it.

Finally the modified search.html page located in the _templates directory overrides
the page created by Sphinx.