Fixed wrong prior distribution in LinearLeastSquaresCalibration and NonLinearLeastSquaresCalibration
In the case of the least squares calibration algorithms, the prior is noninformative, it means it is equal to the Lebesgue measure instead of a Dirac distribution. This measure is emulated using a Normal distribution with a variance equal to SpecFunc.MaxScalar. Fixed issue #1594.

When the prior and posterior distributions are very different, the graphs produced by CalibrationResult are almost useless, so the relevant methods have been modified to ensure that both the posterior distribution and the candidate are always visible, even if it requires to truncate the graph of the prior pdf. The fix is not perfect as the graph remains difficult to read when the candidate is far away from the posterior distribution, but I can't see a way to fix it if both the candidate and the posterior distribution have to be on the same graph...