Improved FittingTest::BestModelLilliefors
Now a first pass is done using the biased Kolmogorov p-value to sort the proposed factories. Then, as soon as a given factory leads to a distribution with an unbiased p-value greater than the biased p-value of another factory, this last one can be disgarded. It allows for massive savings in computation time while remaining exact.