Implemeting KroneckerCovarianceModel
This model allows building multivariate covariance models as kronecker
product of a spatial covariance structure & a correlation model.

We add some validation when we compare expectation of the model
with evaluation built by hand

Also we add unit tests. We check also that exotic situations (isotropic
model) might be set as correlation model