Improved distribution arithmetic
When the length range of the antecedent is much smaller than 1, the computed PDF of a composite distribution was computed as zero for all values of the argument. Now, the absolute precision of the underlying solver is adapted to the range. The logic behind Distribution::Pow has also been improved.