Improved ConditionalDistribution
Now it is possible to select the discretization method for the continuous part using ResourceMap. It can be either "GaussProduct" (default), or "QMC" (using Sobol sequences) or any other value for MonteCarlo. It is much more accurate when a part of the conditioned parameters is not regular (i.e. influence the range of the distribution)