Improved Solver stability
The Bisection, Brent and Secant solvers implemented the change of sign test in a way sensitive to underflow and overflow. Now it uses a test based on individual signs instead of a product. This fixes issue #1805.
The convergence rule was not the same in the different solvers: it has been changed to match the rule in the asserts.