Implement CovarianceModel::computeCrossCovariance
This allows to compute the covariance interaction between X & Y samples.
or X/p where X is a Sample, p a Point

We get the following signatures:
 - computeCrossCovariance(Sample&, Sample&)
 - computeCrossCovariance(Sample&, Point&)
 - computeCrossCovariance(Point&, Sample&)

The last method is less efficient as it relies on the second one +
transpose. We need to improve it later but probably it will not be
intensively used.

These methods were implemented (but private) in KrigingResult for Kriging
variance purposes. Now we implement the method in the CovarianceModel
class as the service is mandatory for some applications such as HSIC

We also implement unit tests in CovarianceModel_std tests and reduce
time for continuous integration purposes

cosmetics