Cobyla: Adaptive rho initialization
The idea is to mimic nlopt logic to set rhobeg/rhoend according to a fraction of problem bounds:
https://github.com/stevengj/nlopt/blob/master/src/api/options.c#L920
https://github.com/stevengj/nlopt/blob/master/src/algs/cobyla/cobyla.c#L219
It helps reducing the number of iteration in some cases, but in other it
will increase, so testing is welcome.

When bounds are defined, rhobeg is set to a fraction of the interval,
and rhoend is relative to rhobeg instead of the fixed values set with
setRhoBeg/setMaximumAbsoluteError.

Some classes (Whittle,BoxCox,Trapezoidal) rely on the fixed behavior,
so it must be disabled, at least if bounds are defined.

It becomes the default and setUseAdaptativeRho accessor or
key Cobyla-DefaultUseAdaptiveRho

Closes #1876