Add Pagmo optimization
This class exposes bio-inspired and evolutionary global optimization algorithms
from the `Pagmo <https://esa.github.io/pagmo2/>`_ library.

These algorithms start from an initial population and make it evolve to obtain
a final population after a defined number of generations.

A few of these algorithms allow for multi-objective optimization, and in that
case the result is not the best point among the final population but a set of
dominant points: a pareto front.

A few algorithms allow for batch evaluation of the populations;
this is reflected in the API with a setBlockSize accessor.