Prevent premature garbage collection of interfaces
Fixes a bug where for example
ot.ComparisonOperator(ot.Less()).getImplementation()
segfaults.

ComparisonOperator and Less could be respectively replaced by any TypedObjectInterface and Implementation.

The reason of the bug was that SWIG could not know that the output of getImplementation() points to an attribute of the TypedObjectInterface.
As a result, after executing the line, Python garbage collected the TypedInterfaceObject and the pointer that was created by getImplementation() became invalid.

The trick is to give the output of getImplementation() in Python an attribute that references the original TypedObjectInterface to prevent its garbage collection.
http://www.swig.org/Doc4.0/Python.html#Python_memory_management_member_variables