Add FireSatellite use-case
PR proposes a multidisciplinary aerospace use case that is the design of a Fire Satellite. This test case is quite popular in the literature (see Mahadevan's papers).

It is a 9 variables use case that is composed of 3 disciplines (orbit, power and attitude) that exchange some coupling variables to compute three outputs of interest (total torque of the satellite, total power and area of the solar array). A fixed point iteration algorithm has been coded in order to solve the multidisciplinary analysis (the tolerance of this algorithm as well as the maximal number of iterations are options of the class).

Two dozens of parameters (deterministic) can be tuned to define different variants of the usecase. This test case can be used for reliability analysis as well as sensitivity analysis.

The reference paper is Sankararaman and Mahadevan (2012) Likelihood-Based Approach to Multidisciplinary Analysis Under Uncertainty, Journal of Mechanical Design (134)031008-1