Added examples
Now the use of univariate functions and basis are illustrated. It also covers the mechanisms allowing to increase the input dimension or the output dimension of a function are illustrated, as well as the tensorization mechanisms.

The example of multivariate, multidimensional basis creation is taken from the description of PR#2212