Improved Triangular and Trapezoidal
The exact expression of the characteristic functions of these two distributions are ill conditioned for arguments of small magnitude. The threshold defining the switch between these formulas and their series expansion were too small, and there was an additional bug for the Trapezoidal case.
The cutoff value between the series expansion and the full expression of the characteristic function was based on an even power of the series instead of an even power, leading to wrong use of the series in case of vanishing odd powers due to symmetry.
Fixes #2216.