Fixed LARS
The constant function, which has the index 0 in any orthogonal basis by convention, must be accepted first in the sequential exploration otherwise the correlation computation through a matrix/vector product is wrong. One easy way to enforce this is to deactivate the selection step based on correlation at the first iteration.