Introduce new methods in Matrix
Until now, to handle diagonal, we had :
 - Matrix::setDiagonal(Point&, int)
 - Matrix::getDiagonal(int ) --> Matrix

We implementt :
 - Matrix::setDiagonal(Scalar, int)
 - Matrix::setDiagonal(Matrix&, int)
 - Matrix::getDiagonalAsPoint(int) --> Point

Very often, in applications, we need to set a diagonal with the same
value. Thus instead of systematically create a Point containing the same
value, we implement this method. This explains the first feature

The other ones help to have a "symmetry" between getter end setter

Implement Matrix::getDiagonalAsPoint & Martix::setDiagonal(Matrix&)